python/requirements.txt
flask
flask_cors
requests
py-postgresql>=1.2.2
pytz
python-dotenv
